Bug fixes around temperature sensors and code reduction
Implemented DS18B20 presence detection per spec
Implemented DS18B20 "general case" crc in anticipation of applying crc to scratchpad  read for temperature measurements
Changed uint8_t to int in numerous places
Fixed bugs in DS18B20 Config messages
Bug fixing around degree symbol in MQTT temperature payload
Removed spaces in IOControl temperature fields

+++ NetworkModule/httpd.h
@@ -77,7 +77,7 @@ void int2hex(uint8_t i);
 
 void HttpDInit(void);
 void HttpDCall(uint8_t* pBuffer, uint16_t nBytes, struct tHttpD* pSocket);
-void parse_POST_string(uint8_t curr_ParseCmd, uint8_t num_chars);
+void parse_POST_string(uint8_t curr_ParseCmd, int num_chars);
 void parse_POST_address(uint8_t curr_ParseCmd, uint8_t curr_ParseNum);
 void parse_POST_port(uint8_t curr_ParseCmd, uint8_t curr_ParseNum);
 void parse_POST_MAC(uint8_t curr_ParseCmd);
